revise the admin form
