--- app/group-admin/register/map.tsx
@@ -1,384 +1,0 @@
-/* eslint-disable import/no-unused-modules */
-import React, { useState, useEffect } from 'react';
-import { StyleSheet, View, Alert, Platform, ScrollView as RNScrollView } from 'react-native';
-import { useRouter, useLocalSearchParams } from 'expo-router';
-import { ThemedView } from '@/components/ThemedView';
-import { ThemedText } from '@/components/ThemedText';
-import { Button } from '@/components/Button';
-import UniversalMap from '@/components/UniversalMap';
-import { useThemeColor } from '@/hooks/useThemeColor';
-import { Spacing, BorderRadius } from '@/constants/Design';
-import { Config } from '@/constants/Config';
-import { withOpacity } from '@/utils/colorUtils';
-import AppHeader from '@/components/AppHeader';
-import * as Location from 'expo-location';
-
-export default function GroupAdminRegisterMap() {
-  const router = useRouter();
-  const { formData, role, roleTitle, roleIcon } = useLocalSearchParams();
-  const roleIconSafe = typeof roleIcon === 'string' ? roleIcon : Array.isArray(roleIcon) ? roleIcon[0] : '📍';
-  const roleTitleSafe = typeof roleTitle === 'string' ? roleTitle : Array.isArray(roleTitle) ? roleTitle[0] : '';
-  const [location, setLocation] = useState<{ latitude: number; longitude: number; address?: string } | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-  const [locError, setLocError] = useState<string | null>(null);
-
-  useEffect(() => {
-    (async () => {
-      try {
-        const { status } = await Location.requestForegroundPermissionsAsync();
-        if (status !== 'granted') {
-          setLocError('مجوز دسترسی به موقعیت داده نشد. لطفاً به صورت دستی مکان را انتخاب کنید.');
-          return;
-        }
-        const pos = await Location.getCurrentPositionAsync({ accuracy: Location.Accuracy.Balanced });
-        setLocation({ latitude: pos.coords.latitude, longitude: pos.coords.longitude });
-      } catch {
-        setLocError('امکان دریافت موقعیت فعلی وجود ندارد.');
-      }
-    })();
-  }, []);
-
-  const primaryColor = useThemeColor({}, 'primary');
-  const successColor = useThemeColor({}, 'success');
-  const errorColor = useThemeColor({}, 'error');
-
-  const handleFinalize = async () => {
-    if (!location) {
-      setError('لطفاً روی نقشه موقعیت را انتخاب کنید');
-      Alert.alert('خطا', 'لطفاً روی نقشه موقعیت را انتخاب کنید');
-      return;
-    }
-
-    setError('');
-    setLoading(true);
-
-    // Navigate to confirmation page
-    setTimeout(() => {
-      router.push({
-        pathname: '/group-admin/register/confirm',
-        params: {
-          formData,
-          role,
-          roleTitle,
-          roleIcon,
-          location: JSON.stringify(location),
-        },
-      });
-      setLoading(false);
-    }, 500);
-  };
-
-  const handleLocationSelect = (loc: { latitude: number; longitude: number }) => {
-    setLocation(loc);
-    setError('');
-  };
-
-  const ProgressBar = () => (
-    <View style={styles.progressContainer}>
-      <View style={[styles.progressStep, { backgroundColor: successColor }]}>
-        <ThemedText style={styles.progressText}>✓</ThemedText>
-      </View>
-      <View style={[styles.progressLine, { backgroundColor: successColor }]} />
-      <View style={[styles.progressStep, { backgroundColor: successColor }]}>
-        <ThemedText style={styles.progressText}>✓</ThemedText>
-      </View>
-      <View style={[styles.progressLine, { backgroundColor: primaryColor }]} />
-      <View style={[styles.progressStep, { backgroundColor: primaryColor }]}>
-        <ThemedText style={styles.progressText}>۳</ThemedText>
-      </View>
-    </View>
-  );
-
-  const INSET_BEHAVIOR: any = 'always';
-  const ANDROID_OVERSCROLL: any = Platform.OS === 'android' ? 'always' : undefined;
-
-  return (
-    <ThemedView type="container" style={styles.container}>
-      <AppHeader title="انتخاب موقعیت جغرافیایی" subtitle={`برای ${roleTitleSafe}`} />
-
-      {/* Top action bar so user can confirm without scrolling */}
-      <View style={[styles.topBar, { backgroundColor: withOpacity(primaryColor, 8) }]}>
-        <Button
-          title="تأیید و ادامه"
-          onPress={handleFinalize}
-          loading={loading}
-          disabled={!location}
-          size="small"
-          variant="success"
-        />
-      </View>
-
-      <RNScrollView
-        style={{ flex: 1 }}
-        keyboardShouldPersistTaps="handled"
-        keyboardDismissMode={Platform.OS === 'ios' ? 'on-drag' : 'none'}
-        contentInsetAdjustmentBehavior={INSET_BEHAVIOR}
-        nestedScrollEnabled
-        overScrollMode={ANDROID_OVERSCROLL}
-        removeClippedSubviews={false}
-        scrollEnabled
-        showsVerticalScrollIndicator={false}
-        contentContainerStyle={{ flexGrow: 1, paddingHorizontal: Spacing.xl, paddingBottom: Spacing['4xl'] }}
-      >
-        {/* Progress Bar */}
-        <ProgressBar />
-
-        {/* Header with Role Badge */}
-        <View style={styles.header}>
-          <View style={[styles.roleIconContainer, { backgroundColor: withOpacity(primaryColor, 20) }]}>
-            <ThemedText style={styles.roleIcon}>{roleIconSafe}</ThemedText>
-          </View>
-          <ThemedText type="heading2" center style={styles.title}>
-            انتخاب موقعیت جغرافیایی
-          </ThemedText>
-          <ThemedText type="body" center style={styles.subtitle}>
-            موقعیت {roleTitleSafe} را روی نقشه مشخص کنید
-          </ThemedText>
-        </View>
-
-        {/* Group Admin Badge */}
-        <ThemedView type="card" style={[styles.badgeCard, { backgroundColor: withOpacity(primaryColor, 10), borderColor: withOpacity(primaryColor, 20) }]}>
-          <View style={styles.badgeContent}>
-            <ThemedText style={styles.badgeIcon}>👥</ThemedText>
-            <ThemedText type="caption" style={styles.badgeText}>
-              ثبت‌نام توسط نماینده گروه - این موقعیت تحت پوشش گروه شما ثبت می‌شود
-            </ThemedText>
-          </View>
-        </ThemedView>
-
-        {/* Map Section */}
-        <ThemedView type="card" style={styles.mapCard}>
-          <ThemedText type="heading3" style={styles.mapTitle}>
-            انتخاب آدرس روی نقشه
-          </ThemedText>
-
-          <View style={styles.mapContainer} onStartShouldSetResponderCapture={() => true}>
-            <UniversalMap
-              location={location}
-              onLocationSelect={handleLocationSelect}
-              mapType="standard"
-              zoom={0.01}
-              showControls={true}
-            />
-          </View>
-
-          {location && (
-            <ThemedView type="surface" style={styles.locationInfo}>
-              <View style={styles.locationHeader}>
-                <ThemedText style={styles.locationIcon}>📍</ThemedText>
-                <ThemedText type="body" weight="medium" style={styles.locationTitle}>
-                  موقعیت انتخاب شده
-                </ThemedText>
-              </View>
-              <ThemedText type="caption" style={styles.coordinates}>
-                عرض جغرافیایی: {location.latitude.toFixed(6)}
-              </ThemedText>
-              <ThemedText type="caption" style={styles.coordinates}>
-                طول جغرافیایی: {location.longitude.toFixed(6)}
-              </ThemedText>
-              {location.address && (
-                <ThemedText type="body" style={styles.address}>
-                  آدرس: {location.address}
-                </ThemedText>
-              )}
-            </ThemedView>
-          )}
-
-          {/* Error State */}
-          {error && (
-            <ThemedView style={[styles.errorContainer, { backgroundColor: withOpacity(errorColor, 10), borderColor: withOpacity(errorColor, 20) }]}>
-              <ThemedText type="caption" style={[styles.errorText, { color: errorColor }]}>
-                ⚠️ {error}
-              </ThemedText>
-            </ThemedView>
-          )}
-
-          {/* Location Error State */}
-          {locError && (
-            <ThemedView style={[styles.errorContainer, { backgroundColor: withOpacity(errorColor, 10), borderColor: withOpacity(errorColor, 20) }]}>
-              <ThemedText type="caption" style={[styles.errorText, { color: errorColor }]}>
-                ⚠️ {locError}
-              </ThemedText>
-            </ThemedView>
-          )}
-        </ThemedView>
-
-        {/* Instructions */}
-        <ThemedView type="card" style={[styles.instructionsCard, { backgroundColor: withOpacity(primaryColor, 5), borderColor: withOpacity(primaryColor, 20) }]}>
-          <ThemedText type="body" weight="medium" style={[styles.instructionsTitle]}>
-            راهنمای استفاده از نقشه:
-          </ThemedText>
-          <ThemedText type="caption" style={styles.instructionText}>
-            • روی نقشه کلیک کنید تا موقعیت انتخاب شود
-          </ThemedText>
-          <ThemedText type="caption" style={styles.instructionText}>
-            • موقعیت انتخابی در منطقه تحت پوشش گروه شما ثبت می‌شود
-          </ThemedText>
-          <ThemedText type="caption" style={styles.instructionText}>
-            • اطمینان حاصل کنید که آدرس صحیح است
-          </ThemedText>
-        </ThemedView>
-
-        {/* Action Buttons */}
-        <View style={styles.buttonContainer}>
-          <Button
-            title="بازگشت"
-            onPress={() => router.back()}
-            variant="outline"
-            style={styles.backButton}
-          />
-          <Button
-            title="تأیید و ادامه"
-            onPress={handleFinalize}
-            loading={loading}
-            disabled={!location}
-            style={styles.continueButton}
-            icon={<ThemedText>✓</ThemedText>}
-          />
-        </View>
-      </RNScrollView>
-    </ThemedView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  topBar: {
-    paddingHorizontal: Spacing.xl,
-    paddingTop: Spacing.md,
-    paddingBottom: Spacing.sm,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-    borderColor: 'rgba(0,0,0,0.06)',
-  },
-  progressContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginBottom: Spacing['3xl'],
-    paddingHorizontal: Spacing.xl,
-  },
-  progressStep: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  progressLine: {
-    flex: 1,
-    height: 2,
-    marginHorizontal: Spacing.sm,
-  },
-  progressText: {
-    color: 'white',
-    fontWeight: 'bold',
-  },
-  header: {
-    alignItems: 'center',
-    marginBottom: Spacing.xl,
-  },
-  roleIconContainer: {
-    width: 80,
-    height: 80,
-    borderRadius: 40,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginBottom: Spacing.lg,
-  },
-  roleIcon: {
-    fontSize: 36,
-  },
-  title: {
-    marginBottom: Spacing.sm,
-  },
-  subtitle: {
-    opacity: 0.7,
-    textAlign: 'center',
-  },
-  badgeCard: {
-    marginBottom: Spacing.xl,
-    borderWidth: 1,
-  },
-  badgeContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  badgeIcon: {
-    fontSize: 24,
-    marginLeft: Spacing.md,
-  },
-  badgeText: {
-    flex: 1,
-    opacity: 0.8,
-  },
-  mapCard: {
-    marginBottom: Spacing.xl,
-  },
-  mapTitle: {
-    marginBottom: Spacing.lg,
-  },
-  mapContainer: {
-    height: 300,
-    borderRadius: 12,
-    overflow: 'hidden',
-    marginBottom: Spacing.md,
-  },
-  locationInfo: {
-    padding: Spacing.md,
-    borderRadius: 8,
-    marginBottom: Spacing.md,
-  },
-  locationHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: Spacing.sm,
-  },
-  locationIcon: {
-    fontSize: 20,
-    marginLeft: Spacing.sm,
-  },
-  locationTitle: {
-    flex: 1,
-  },
-  coordinates: {
-    opacity: 0.7,
-    marginBottom: Spacing.xs,
-  },
-  address: {
-    marginTop: Spacing.sm,
-  },
-  errorContainer: {
-    padding: Spacing.md,
-    borderRadius: 8,
-    borderWidth: 1,
-  },
-  errorText: {
-    textAlign: 'center',
-  },
-  instructionsCard: {
-    marginBottom: Spacing.xl,
-    borderWidth: 1,
-  },
-  instructionsTitle: {
-    marginBottom: Spacing.md,
-    color: '#2E7D32',
-  },
-  instructionText: {
-    marginBottom: Spacing.xs,
-    opacity: 0.8,
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: Spacing['4xl'],
-  },
-  backButton: {
-    flex: 0.45,
-  },
-  continueButton: {
-    flex: 0.45,
-  },
-});
